Rewrite
Removed the knowledge about actual changes from TreeController, now it
only signals when there is some change.

TreeDataStore resets the sections root tree controller and compares
actual vs expected. If there is a difference it makes changes to its
cache and it notifies the tree control in its turn.

Simple!